# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganizationCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 255,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "OrganizationCreateRequest",
        "type": "object"
      },
...
      "OrganizationListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrganizationResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "OrganizationListResponse",
        "type": "object"
      },
...
      "OrganizationResponse": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/OrganizationCapabilitiesResponse"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "current_role": {
            "enum": [
              "admin",
              "manager",
              "viewer"
            ],
            "title": "Current Role",
            "type": "string"
          },
          "enforce_mfa": {
            "title": "Enforce Mfa",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "uuid": {
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "uuid",
          "name",
          "enforce_mfa",
          "current_role",
          "capabilities",
          "created_at",
          "updated_at"
        ],
        "title": "OrganizationResponse",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations": {
      "get": {
        "operationId": "list_organizations_api_v1_organizations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Organizations",
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "create_organization_api_v1_organizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Organization",
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/{organization_uuid}": {
      "delete": {
        "operationId": "delete_organization_api_v1_organizations__organization_uuid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "organization_uuid",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Organization",
        "tags": [
          "organizations"
        ]
      },
      "get": {
        "operationId": "get_organization_api_v1_organizations__organization_uuid__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_uuid",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OrganizationLoginDetailResponse"
                    },
                    {
                      "$ref": "#/components/schemas/OrganizationIntegrationDetailResponse"
                    }
                  ],
                  "title": "Response Get Organization Api V1 Organizations  Organization Uuid  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Organization",
        "tags": [
          "organizations"
        ]
      },
      "patch": {
        "operationId": "update_organization_api_v1_organizations__organization_uuid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "organization_uuid",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationLoginDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Update Organization",
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/{organization_uuid}/billing-transfers": {
      "post": {
        "operationId": "transfer_billing_api_v1_organizations__organization_uuid__billing_transfers_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_uuid",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/rentivo__api__schemas__organizations__BillingTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingTransferResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfer Billing",
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/{organization_uuid}/invites": {
      "post": {
        "operationId": "create_invite_api_v1_organizations__organization_uuid__invites_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_uuid",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationInviteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationInviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invite",
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/api/v1/organizations/{organization_uuid}/mfa-policy": {
      "put": {
        "operationId": "update_mfa_policy_api_v1_organizations__organization_uuid__mfa_policy_put",
        "parameters": [
          {
            "in": "path",
            "name": "organization_uuid",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationMFAPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMFAPolicyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Mfa Policy",
        "tags": [
          "organizations"
        ]
      }
    },
...
  }
}